Added favicon.ico
A quick and simple favicon.ico based on the nice design at
css3test.com, just the curly brackets on pink background.

Icon will automatically display in all versions of Opera, Firefox, IE,
Chrome, Safari etc. The icon is 32x32 pixels which is used in IE10 and
all other browsers auto-scale the icon down  to 16x16. Just drop it in
the root folder.