Switched 'revert' value of 'all' to CSS-wide 'revert' keyword
A previous PR adding revert only added it as a value of ‘all’, so only
applies if that property is supported. I’ve changed it to the real
change, which is adding ‘revert’ as a CSS-wide keyword. I switched out
the last test from ‘width’ to ‘all’ to make sure it works on that
property.