Test unicode-range as a descriptor, not a property.
This implements support for testing whether a value of an @font-face
descriptor is supported.  I've added the code to do this in supports.js,
although it's not clear to me whether I should be adding it there, or
whether you'd rather leave supports.js untouched and add this code
outside of it.

The test for testing descriptors works somewhat differently from other
properties because there isn't any better object model support that's
interoperable.  I've tested locally in Firefox and Chrome that supported
descriptors pass and unsupported descriptors fail.

I didn't add support for prefixing in the tests of descriptor support.
It didn't seem necessary for the unicode-range test, it's extra work,
and I tend to think it's also harmful to promote prefixed support.

This flips the unicode-range tests from red to green on both Firefox and
Chrome.  I haven't tested elsewhere.