Don't test for prefixed support.
This patch removes the tests for prefixed support so that only support
for unprefixed implementation of the features is tested.

I realize it's not obvious that this change is the right thing to do,
but I think that given current thoughts on prefixing and its influence
on the Web, I think this change is worth considering.

Widespread browser support for prefixed CSS has led to large amounts of
browser-specific CSS on the Web; promoting support for unprefixed CSS
helps encourage browsers to behave in a way that encourages
cross-browser content on the Web.